SymbolicCRN

Documentation for SymbolicCRN.

SymbolicCRN.tpoly

AbstractAlgebra.matrix

SymbolicCRN.Diagonal

SymbolicCRN.Jacobian

SymbolicCRN.cone_positivenullspace

SymbolicCRN.cone_positiveorthant

SymbolicCRN.cone_vectorspace

SymbolicCRN.dissect

SymbolicCRN.integermultiple

SymbolicCRN.tpoly — Type

Structure of a tpolynomial T. Vector p of coefficients of T, with v[1] != 0 (nonzero independent term) and integer mindeg idicating the smallest degree of T. mindeg can be negative, we have, with n = length(v)-1,

 T = x^mindeg * sum( v .* [1, x,..., x^n])

AbstractAlgebra.matrix — Method

matrix(R, V::AbstractVector{T}) where {T}

Given a Julia vector V of entries, construct the corresponding AbstractAlgebra.jl one-column matrix over the given ring R, assuming all the entries can be coerced into R.

SymbolicCRN.Diagonal — Method

Diagonal(R, V::AbstractVector{T}) where {T}

Given a Julia vector V of entries, construct the corresponding AbstractAlgebra.jl diagonal matrix over the given ring R, assuming all the entries can be coerced into R.

SymbolicCRN.Jacobian — Function

Jacobian(R, M)
Jacobian(R, M, vars=1:length(gens(R)))

Returns the Jacobian matrix of a one-column matrix of polynomials M with respect to the generators of R indexed by vars. When vars is omitted all the generators of R are used.

SymbolicCRN.cone_positivenullspace — Method

cone_positivenullspace(N::AbstractMatrix{T}) where {T<:Integer}

Return the cone (Polymake big object) intersection of the nonnegative orthant and the nullspace of N.

SymbolicCRN.cone_positiveorthant — Method

cone_positiveorthant(n)

Return the cone (Polymake big object) corresponding to the nonnegative orthant of R^n.

SymbolicCRN.cone_vectorspace — Method

cone_vectorspace(M::AbstractMatrix{T}) where {T <: Integer}

Return the cone (Polymake big object) corresponding to vector space generated by the columns of M.

SymbolicCRN.dissect — Method

dissect(p::MPolyElem)

Iterator for the coefficients and exponent vectors of the given polynomial.

SymbolicCRN.integermultiple — Method

integermultiple(A)

Given an array of rational numbers A, returns a multiple λA with integer entries, where λ=abs(lcm(denominator.(A))) is the minimal integer with this property.