SymbolicCRN

Documentation for SymbolicCRN.

SymbolicCRN.PolyPolyt

SymbolicCRN.tpoly

AbstractAlgebra.matrix

SymbolicCRN.Jacobian

SymbolicCRN.Jacobian_cp

SymbolicCRN.Newtonpolytope

SymbolicCRN.coeff_charpoly

SymbolicCRN.cone_positivenullspace

SymbolicCRN.cone_positiveorthant

SymbolicCRN.cone_vectorspace

SymbolicCRN.diagonal

SymbolicCRN.dissect

SymbolicCRN.integermultiple

SymbolicCRN.save_polypolyt

SymbolicCRN.vertex_index

SymbolicCRN.vertex_point_map

SymbolicCRN.PolyPolyt — Type

struct PolyPolyt{T}

A polynomial p, its Newton polytope Newtonpolyt (as Polymake big object, with the vertices computed) and the map vertices -> points vertex_point_map (vector storing the position of vertex in polynomial p).

Methods to create a PolyPolyt:

PolyPolyt(p::MPolyElem) # Requires computing the vertices of the Newton polytope.
PolyPolyt(p::MPolyElem, vertices, ver_pt_map) # Loads vertices and ver-to-pt map.
PolyPolyt(p::MPolyElem, name) # Loads them from files generated by save_polypolyt(name::String, pp)

SymbolicCRN.tpoly — Type

Structure of a tpolynomial P. Vector p of coefficients of P, with v[1] != 0 (nonzero independent term) and integer mindeg indicating the smallest degree of P. mindeg can be negative, we have, with n = length(v)-1,

     poly = t^(P.mindeg) * sum( P.p .* [1, t,..., t^n])

AbstractAlgebra.matrix — Method

matrix(R, V)

Given a Julia vector V of entries, construct the corresponding AbstractAlgebra.jl one-column matrix over the given ring R, assuming all the entries can be coerced into R.

SymbolicCRN.Jacobian — Function

Jacobian(R, F)
Jacobian(R, F, vars)

Returns the Jacobian matrix of a one-column matrix of polynomials F with respect to the generators of R indexed by vars. When vars is omitted all the generators of R are used.

Examples

julia> using Nemo

julia> R, vars = PolynomialRing(ZZ, vcat(["k$i" for i in 1:5], ["x$i" for i in 1:4]));

julia> F = [vars[1]*vars[2]*vars[6]-vars[8]; vars[3]*vars[9]+2*vars[7]]
[k1*k2*x1-x3]
[ k3*x4+2*x2]

julia> Jacobian(R, F, 6:9)
[k1*k2  0  -1   0]
[    0  2   0  k3]

julia> Jacobian(R, F)
[k2*x1  k1*x1   0  0  0  k1*k2  0  -1   0]
[    0      0  x4  0  0      0  2   0  k3]

SymbolicCRN.Jacobian_cp — Method

Jacobian_cp(N, Y, E, ls, hs)

Returns Nemo's MatElem paramatrising the Jacobian of the dynamical system associated to a biochemical reaction network. We use "convex parameters", see

rn = @reaction_network begin
    k1,  X2 → X1
    k2,  X1 → X2
end k1 k2

N = netstoichmat(rn)
Y = substoichmat(rn)

E = raysof(cone_positivenullspace(N))
W = conservationlaws(N)

SymbolicCRN.Newtonpolytope — Method

Newtonpolytope(pp)

Returns the Newton polytope of a PolyPolyt.

SymbolicCRN.coeff_charpoly — Method

coeff_charpoly(M, i)

Returns the i-th coefficient of the characteristic polynomial of M.

SymbolicCRN.cone_positivenullspace — Method

Return the cone (Polymake big object) intersection of the nonnegative orthant and the nullspace of N.

SymbolicCRN.cone_positiveorthant — Method

cone_positiveorthant(n)

Return the cone (Polymake big object) corresponding to the nonnegative orthant of R^n.

SymbolicCRN.cone_vectorspace — Method

cone_vectorspace(M)

Return the cone (Polymake big object) corresponding to vector space generated by the columns of M.

SymbolicCRN.diagonal — Method

diagonal(R, V)

Given a Julia vector V of entries, construct the corresponding AbstractAlgebra.jl diagonal matrix over the given ring R, assuming all the entries can be coerced into R.

SymbolicCRN.dissect — Method

dissect(p)

Iterator for the coefficients and exponent vectors of the given polynomial.

SymbolicCRN.integermultiple — Function

Given an array of rational numbers A, returns a multiple λA whose entries are all integers, where λ=abs(lcm(denominator.(A))) is the minimal integer with this property.

SymbolicCRN.save_polypolyt — Method

save_polypolyt(name, pp)

Save vertices of pp::PolyPolyt in file name * ver_sufix and ver-to-pt map in name * map_sufix

SymbolicCRN.vertex_index — Method

vertex_index(pp, i)

Given a vertex i of a PolyPolyt, returns the position of i in ver-to-pt map. That is, vertex_point_map(pp)[vertex_index(pp, i)] == i. Or more interestingly,

verticesof(Newtonpolytope(pp))[:, vertex_index(pp, i)] == collect(exponent_vectors(pp.p))[:,i]

SymbolicCRN.vertex_point_map — Method

vertex_point_map(pp)

Returns the ver-to-pt map of a PolyPolyt.